refactor: introduce commands package
